Restrict Pin View and Open View to active CPP debug context.
That is if any CPP application is under debug. Because Pin View is never
enabled other than CPP application debug session.

This also includes Action to Command migration.

Fixes https://github.com/eclipse-cdt/cdt/issues/1048
